Mejorado el main

+++ TFGapp/TFGapp/Profesor.cpp
@@ -16,7 +16,7 @@ void Profesor::printProfesor()
 	std::cout << "-------------------" << '\n';
 	std::cout << "Name: " << this->nombreCompleto << '\n';
 	std::cout << "Doctor: " << this->doctor << '\n';
-	for (int i = 0; i < this->grados.size(); i++)
+	for (int i = 0; i < (int)this->grados.size(); i++)
 	{
 		std::cout << "Grado: " << this->grados[i]->getNombre() <<"  nMAX: " << this->nTFG[i] <<  '\n';
 	}
